# oj-forntend/src/api/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "oj-backend",
    "version": "1.0",
    "title": "接口文档"
  },
  "host": "localhost:8101",
...
  "paths": {
...
    "/api/post_favour/": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "doPostFavour",
        "operationId": "doPostFavourUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postFavourAddRequest",
            "description": "postFavourAddRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostFavourAddRequest",
              "originalRef": "PostFavourAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«int»",
              "originalRef": "BaseResponse«int»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post_favour/list/page": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "listFavourPostByPage",
        "operationId": "listFavourPostByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postFavourQueryRequest",
            "description": "postFavourQueryRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostFavourQueryRequest",
              "originalRef": "PostFavourQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»",
              "originalRef": "BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post_favour/my/list/page": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "listMyFavourPostByPage",
        "operationId": "listMyFavourPostByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostQueryRequest",
              "originalRef": "PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»",
              "originalRef": "BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/user/get/login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getLoginUser",
        "operationId": "getLoginUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«LoginUserVO»",
              "originalRef": "BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BaseResponse«LoginUserVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/LoginUserVO",
          "originalRef": "LoginUserVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«LoginUserVO»"
    },
    "BaseResponse«Page«PostVO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page«PostVO»",
          "originalRef": "Page«PostVO»"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Page«PostVO»»"
    },
...
    "BaseResponse«int»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«int»"
    },
...
    "LoginUserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "LoginUserVO"
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean"
        },
        "column": {
          "type": "string"
        }
      },
      "title": "OrderItem"
    },
    "Page«PostVO»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PostVO",
            "originalRef": "PostVO"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«PostVO»"
    },
...
    "PostFavourAddRequest": {
      "type": "object",
      "properties": {
        "postId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostFavourAddRequest"
    },
    "PostFavourQueryRequest": {
      "type": "object",
      "properties": {
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "pageSize": {
          "type": "integer",
          "format": "int64"
        },
        "postQueryRequest": {
          "$ref": "#/definitions/PostQueryRequest",
          "originalRef": "PostQueryRequest"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostFavourQueryRequest"
    },
    "PostQueryRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "favourUserId": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "notId": {
          "type": "integer",
          "format": "int64"
        },
        "orTags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pageSize": {
          "type": "integer",
          "format": "int64"
        },
        "searchText": {
          "type": "string"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostQueryRequest"
    },
...
    "PostVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "hasFavour": {
          "type": "boolean"
        },
        "hasThumb": {
          "type": "boolean"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tagList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "user": {
          "$ref": "#/definitions/UserVO",
          "originalRef": "UserVO"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostVO"
    },
...
    "UserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserVO"
    }
  }
}